pyproject.toml
[project]
name = "ai-delivery-analyzer"
version = "0.1.0"
description = "Agentic requirement analyzer with retrieval, ambiguity detection, and self-reflection."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "AI Academy Capstone", email = "team@example.com" }]
dependencies = [
  "openai>=1.12.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "typer[all]>=0.12.0",
  "rich>=13.7.0",
  "faiss-cpu>=1.7.4",
  "tqdm>=4.66.0",
  "anthropic>=0.26.0",
  "google-generativeai>=0.4.0",
  "numpy>=1.24.0"
]

[project.scripts]
ai-analyze = "cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
